donut-gradio-app/requirements.txt
gradio
transformers
torch
Pillow
sentencepiece